python_app/requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
pymongo>=4.6.0
python-dotenv>=1.0.0
bcrypt>=4.1.0

# Optional: legacy SQLite dev (not used when MONGO_URI is set)
sqlalchemy>=2.0.0
